@@ -1,10 +1,11 @@
-import { Component, Input } from '@angular/core';
+import { Component, inject, Input } from '@angular/core';
 import { PartialMessage, StorageService } from '../../storage.service';
 
 import { FormsModule } from '@angular/forms';
 import { RouterLink } from '@angular/router';
 import { HasAccessToLinkDirective } from '../../../../components/has-access-to-link.directive';
 import { LaddaModule } from 'angular2-ladda';
+import { SweetalertService } from '../../../../services/sweetalert.service';
 
 @Component({
   selector: 'app-storage-list-dt',
@@ -15,5 +16,22 @@ import { LaddaModule } from 'angular2-ladda';
 export class StorageListDtComponent {
   @Input({ required: true }) message!: PartialMessage;
 
-  constructor(public storageService: StorageService) {}
+  protected storageService: StorageService = inject(StorageService);
+  private sweetAlert: SweetalertService = inject(SweetalertService);
+
+  moveMessage(): void {
+    this.sweetAlert
+      .Warning({
+        title: 'Move Message State',
+        text: 'The message might still be processing in the background. Are you sure you want to move it to Error?',
+        confirmButtonText: 'Move to Error',
+        cancelButtonText: 'Cancel',
+        showCancelButton: true,
+      })
+      .then((value) => {
+        if (value.isConfirmed) {
+          this.storageService.moveMessage(this.message);
+        }
+      });
+  }
 }
